requirements.txt
awscli==1.11.97
boto3==1.4.4
botocore==1.5.60
colorama==0.3.7
dj-static==0.0.6
Django==1.11.2
django-cors-headers==2.1.0
django-filter==1.0.4
django-storages==1.5.2
djangorestframework==3.6.3
docutils==0.13.1
jmespath==0.9.3
Markdown==2.6.8
olefile==0.44
Pillow==4.1.1
pyasn1==0.2.3
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
rsa==3.4.2
s3transfer==0.1.10
six==1.10.0
static3==0.7.0
watchtower==0.3.6
